knowledgebase_rag/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.20.0
requests==2.32.3
unstructured==0.14.10
pinecone-client==4.1.2
beautifulsoup4==4.12.3
